pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hwtester"
version = "0.1.0"
description = "Hardware tester CLI - Control relays and log DUT serial output"
requires-python = ">=3.9"
license = {text = "MIT"}
keywords = ["hardware", "testing", "serial", "relay", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Hardware",
]
dependencies = [
    "pyserial>=3.5",
    "tomli>=2.0; python_version < '3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
hwtester = "hwtester.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["hwtester*"]
